requirements.txt
#Requirements for this application are:

altair==4.2.0
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
branca==0.6.0
cachetools==5.2.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==22.3.4
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==5.1.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
folium==0.13.0
future==0.18.2
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.29
greenlet==2.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==3.0.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.1.1
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.23.5
oauthlib==3.1.0
packaging==21.3
pandas==1.5.2
pexpect==4.6.0
Pillow==9.3.0
probableparsing==0.0.1
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydeck==0.8.0
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyparted==3.11.2
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rich==12.6.0
SecretStorage==2.3.1
semver==2.13.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sos==4.4
SQLAlchemy==1.4.44
ssh-import-id==5.10
streamlit==1.15.1
streamlit-aggrid==0.3.3
streamlit-folium==0.7.0
systemd-python==234
toml==0.10.2
toolz==0.12.0
tornado==6.2
Twisted==18.9.0
typing-extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ubuntu-advantage-tools==27.11.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usaddress==0.5.10
validators==0.20.0
wadllib==1.3.3
WALinuxAgent==2.2.46
watchdog==2.1.9
zipp==1.0.0
zope.interface==4.7.1
